contactForm functionnal + add controllers
