+++ app/views/contact.ejs
@@ -7,22 +7,29 @@
       <main>
         <div class="contact">
           <h1>CONTACT</h1>
-          <form class="contact-form" action="/contact" method="POST">      
+          <% if (alertMessage) { %>
+            <div class="alert"><%= alertMessage %></div>
+          <% } %>
+          
+          <% if (successMessage) { %>
+            <div class="success"><%= successMessage %></div>
+          <% } %>
+          <form class="contact-form" action="/contact" method="post">      
             <label>
               Nom:
-              <input type="text" name="Nom">
+              <input type="text" name="lastname" required>
             </label>      
             <label>
               Prénom:
-              <input type="text" name="Prénom">
+              <input type="text" name="firstname" required>
             </label>      
             <label>
               Email:
-              <input type="email" name="email">
+              <input type="email" name="email" required>
             </label>      
             <label>
               Votre message:
-              <textarea name="message"></textarea>
+              <textarea name="message" required></textarea>
             </label>      
             <button type="submit">Envoyer</button>
           </form>
